--- a/src/array.c
+++ b/src/array.c
@@ -124,7 +124,6 @@ static jl_array_t *_new_array_(jl_value_t *atype, uint32_t ndims, size_t *dims,
     a->flags.ptrarray = !isunboxed;
     a->elsize = elsz;
     a->flags.isshared = 0;
-    a->flags.isaligned = 1;
     a->offset = 0;
     if (ndims == 1) {
         a->nrows = nel;
@@ -187,7 +186,6 @@ JL_DLLEXPORT jl_array_t *jl_reshape_array(jl_value_t *atype, jl_array_t *data,
     a->flags.ndims = ndims;
     a->offset = 0;
     a->data = NULL;
-    a->flags.isaligned = data->flags.isaligned;
     jl_value_t *el_type = jl_tparam0(atype);
     assert(store_unboxed(el_type) == !data->flags.ptrarray);
     if (!data->flags.ptrarray) {
@@ -247,7 +245,6 @@ JL_DLLEXPORT jl_array_t *jl_string_to_array(jl_value_t *str)
     a->flags.ndims = 1;
     a->offset = 0;
     a->data = jl_string_data(str);
-    a->flags.isaligned = 0;
     a->elsize = 1;
     a->flags.ptrarray = 0;
     jl_array_data_owner(a) = str;
@@ -289,7 +286,6 @@ JL_DLLEXPORT jl_array_t *jl_ptr_to_array_1d(jl_value_t *atype, void *data,
     a->flags.ptrarray = !isunboxed;
     a->flags.ndims = 1;
     a->flags.isshared = 1;
-    a->flags.isaligned = 0;  // TODO: allow passing memalign'd buffers
     if (own_buffer) {
         a->flags.how = 2;
         jl_gc_track_malloced_array(ptls, a, 0);
@@ -345,7 +341,6 @@ JL_DLLEXPORT jl_array_t *jl_ptr_to_array(jl_value_t *atype, void *data,
     a->flags.ndims = ndims;
     a->offset = 0;
     a->flags.isshared = 1;
-    a->flags.isaligned = 0;
     if (own_buffer) {
         a->flags.how = 2;
         jl_gc_track_malloced_array(ptls, a, 0);
@@ -594,7 +589,6 @@ static int NOINLINE array_resize_buffer(jl_array_t *a, size_t newlen)
             a->data = jl_gc_malloc_aligned(nbytes, JL_CACHE_BYTE_ALIGNMENT);
             jl_gc_track_malloced_array(ptls, a, 1);
             a->flags.how = 2;
-            a->flags.isaligned = 1;
         }
         else {
             a->data = jl_gc_alloc_buf(ptls, nbytes);