--- a/src/array.c
+++ b/src/array.c
@@ -102,13 +102,13 @@ static jl_array_t *_new_array_(jl_value_t *atype, uint32_t ndims, size_t *dims,
     }
     else {
         tsz = JL_ARRAY_ALIGN(tsz, JL_CACHE_BYTE_ALIGNMENT); // align whole object
-        data = jl_gc_managed_malloc(tot);
+        data = jl_gc_malloc_aligned(tot, JL_CACHE_BYTE_ALIGNMENT);
         // Allocate the Array **after** allocating the data
         // to make sure the array is still young
         a = (jl_array_t*)jl_gc_alloc(ptls, tsz, atype);
         // No allocation or safepoint allowed after this
         a->flags.how = 2;
-        jl_gc_track_malloced_array(ptls, a);
+        jl_gc_track_malloced_array(ptls, a, 1);
         if (!isunboxed)
             memset(data, 0, tot);
     }
@@ -292,7 +292,7 @@ JL_DLLEXPORT jl_array_t *jl_ptr_to_array_1d(jl_value_t *atype, void *data,
     a->flags.isaligned = 0;  // TODO: allow passing memalign'd buffers
     if (own_buffer) {
         a->flags.how = 2;
-        jl_gc_track_malloced_array(ptls, a);
+        jl_gc_track_malloced_array(ptls, a, 0);
         jl_gc_count_allocd(nel*elsz + (elsz == 1 ? 1 : 0));
     }
     else {
@@ -348,7 +348,7 @@ JL_DLLEXPORT jl_array_t *jl_ptr_to_array(jl_value_t *atype, void *data,
     a->flags.isaligned = 0;
     if (own_buffer) {
         a->flags.how = 2;
-        jl_gc_track_malloced_array(ptls, a);
+        jl_gc_track_malloced_array(ptls, a, 0);
         jl_gc_count_allocd(nel*elsz + (elsz == 1 ? 1 : 0));
     }
     else {
@@ -566,8 +566,7 @@ static int NOINLINE array_resize_buffer(jl_array_t *a, size_t newlen)
     if (a->flags.how == 2) {
         // already malloc'd - use realloc
         char *olddata = (char*)a->data - oldoffsnb;
-        a->data = jl_gc_managed_realloc(olddata, nbytes, oldnbytes,
-                                        a->flags.isaligned, (jl_value_t*)a);
+        a->data = jl_gc_realloc_aligned(olddata, nbytes);
     }
     else if (a->flags.how == 3 && jl_is_string(jl_array_data_owner(a))) {
         // if data is in a String, keep it that way
@@ -592,8 +591,8 @@ static int NOINLINE array_resize_buffer(jl_array_t *a, size_t newlen)
             elsz > 4
 #endif
             ) {
-            a->data = jl_gc_managed_malloc(nbytes);
-            jl_gc_track_malloced_array(ptls, a);
+            a->data = jl_gc_malloc_aligned(nbytes, JL_CACHE_BYTE_ALIGNMENT);
+            jl_gc_track_malloced_array(ptls, a, 1);
             a->flags.how = 2;
             a->flags.isaligned = 1;
         }