--- a/src/array.c
+++ b/src/array.c
@@ -54,7 +54,7 @@ size_t jl_arr_xtralloc_limit = 0;
 #define MAXINTVAL (((size_t)-1)>>1)
 
 static jl_array_t *_new_array_(jl_value_t *atype, uint32_t ndims, size_t *dims,
-                               int isunboxed, int elsz)
+                               int isunboxed, int elsz, int iszeros)
 {
     jl_ptls_t ptls = jl_get_ptls_states();
     size_t i, tot, nel=1;
@@ -97,20 +97,22 @@ static jl_array_t *_new_array_(jl_value_t *atype, uint32_t ndims, size_t *dims,
         // No allocation or safepoint allowed after this
         a->flags.how = 0;
         data = (char*)a + doffs;
-        if (tot > 0 && !isunboxed)
+        if (tot > 0 && (!isunboxed || iszeros))
             memset(data, 0, tot);
     }
     else {
         tsz = JL_ARRAY_ALIGN(tsz, JL_CACHE_BYTE_ALIGNMENT); // align whole object
-        data = jl_gc_malloc_aligned(tot, JL_CACHE_BYTE_ALIGNMENT);
+        if (!isunboxed || iszeros) {
+            data = jl_gc_calloc_aligned(1, tot, JL_CACHE_BYTE_ALIGNMENT);
+        } else {
+            data = jl_gc_malloc_aligned(tot, JL_CACHE_BYTE_ALIGNMENT);
+        }
         // Allocate the Array **after** allocating the data
         // to make sure the array is still young
         a = (jl_array_t*)jl_gc_alloc(ptls, tsz, atype);
         // No allocation or safepoint allowed after this
         a->flags.how = 2;
         jl_gc_track_malloced_array(ptls, a, 1);
-        if (!isunboxed)
-            memset(data, 0, tot);
     }
     a->flags.pooled = tsz <= GC_MAX_SZCLASS;
 
@@ -138,20 +140,20 @@ static jl_array_t *_new_array_(jl_value_t *atype, uint32_t ndims, size_t *dims,
     return a;
 }
 
-static inline jl_array_t *_new_array(jl_value_t *atype, uint32_t ndims, size_t *dims)
+static inline jl_array_t *_new_array(jl_value_t *atype, uint32_t ndims, size_t *dims, int iszeros)
 {
     int isunboxed=0, elsz=sizeof(void*);
     jl_value_t *el_type = jl_tparam0(atype);
     isunboxed = store_unboxed(el_type);
     if (isunboxed)
         elsz = jl_datatype_size(el_type);
-    return _new_array_(atype, ndims, dims, isunboxed, elsz);
+    return _new_array_(atype, ndims, dims, isunboxed, elsz, iszeros);
 }
 
 jl_array_t *jl_new_array_for_deserialization(jl_value_t *atype, uint32_t ndims, size_t *dims,
                                              int isunboxed, int elsz)
 {
-    return _new_array_(atype, ndims, dims, isunboxed, elsz);
+    return _new_array_(atype, ndims, dims, isunboxed, elsz, 0);
 }
 
 #ifndef NDEBUG
@@ -355,35 +357,35 @@ JL_DLLEXPORT jl_array_t *jl_ptr_to_array(jl_value_t *atype, void *data,
     return a;
 }
 
-JL_DLLEXPORT jl_array_t *jl_new_array(jl_value_t *atype, jl_value_t *_dims)
+JL_DLLEXPORT jl_array_t *jl_new_array(jl_value_t *atype, jl_value_t *_dims, int iszeros)
 {
     size_t ndims = jl_nfields(_dims);
     assert(is_ntuple_long(_dims));
-    return _new_array(atype, ndims, (size_t*)_dims);
+    return _new_array(atype, ndims, (size_t*)_dims, iszeros);
 }
 
-JL_DLLEXPORT jl_array_t *jl_alloc_array_1d(jl_value_t *atype, size_t nr)
+JL_DLLEXPORT jl_array_t *jl_alloc_array_1d(jl_value_t *atype, size_t nr, int iszeros)
 {
-    return _new_array(atype, 1, &nr);
+    return _new_array(atype, 1, &nr, iszeros);
 }
 
 JL_DLLEXPORT jl_array_t *jl_alloc_array_2d(jl_value_t *atype, size_t nr,
-                                           size_t nc)
+                                           size_t nc, int iszeros)
 {
     size_t d[2] = {nr, nc};
-    return _new_array(atype, 2, &d[0]);
+    return _new_array(atype, 2, &d[0], iszeros);
 }
 
 JL_DLLEXPORT jl_array_t *jl_alloc_array_3d(jl_value_t *atype, size_t nr,
-                                           size_t nc, size_t z)
+                                           size_t nc, size_t z, int iszeros)
 {
     size_t d[3] = {nr, nc, z};
-    return _new_array(atype, 3, &d[0]);
+    return _new_array(atype, 3, &d[0], iszeros);
 }
 
 JL_DLLEXPORT jl_array_t *jl_pchar_to_array(const char *str, size_t len)
 {
-    jl_array_t *a = jl_alloc_array_1d(jl_array_uint8_type, len);
+    jl_array_t *a = jl_alloc_array_1d(jl_array_uint8_type, len, 0);
     memcpy(a->data, str, len);
     return a;
 }
@@ -427,7 +429,7 @@ JL_DLLEXPORT jl_value_t *jl_cstr_to_string(const char *str)
 
 JL_DLLEXPORT jl_array_t *jl_alloc_vec_any(size_t n)
 {
-    return jl_alloc_array_1d(jl_array_any_type, n);
+    return jl_alloc_array_1d(jl_array_any_type, n, 0);
 }
 
 JL_DLLEXPORT jl_value_t *jl_apply_array_type(jl_value_t *type, size_t dim)
@@ -908,7 +910,7 @@ JL_DLLEXPORT jl_array_t *jl_array_copy(jl_array_t *ary)
 {
     size_t elsz = ary->elsize;
     jl_array_t *new_ary = _new_array_(jl_typeof(ary), jl_array_ndims(ary),
-                                      &ary->nrows, !ary->flags.ptrarray, elsz);
+                                      &ary->nrows, !ary->flags.ptrarray, elsz, 0);
     memcpy(new_ary->data, ary->data, jl_array_len(ary) * elsz);
     return new_ary;
 }