--- a/src/module.c
+++ b/src/module.c
@@ -553,7 +553,7 @@ JL_DLLEXPORT void jl_set_current_module(jl_value_t *m)
 
 JL_DLLEXPORT jl_value_t *jl_module_usings(jl_module_t *m)
 {
-    jl_array_t *a = jl_alloc_array_1d(jl_array_any_type, 0);
+    jl_array_t *a = jl_alloc_array_1d(jl_array_any_type, 0, 0);
     JL_GC_PUSH1(&a);
     for(int i=(int)m->usings.len-1; i >= 0; --i) {
         jl_array_grow_end(a, 1);
@@ -566,7 +566,7 @@ JL_DLLEXPORT jl_value_t *jl_module_usings(jl_module_t *m)
 
 JL_DLLEXPORT jl_value_t *jl_module_names(jl_module_t *m, int all, int imported)
 {
-    jl_array_t *a = jl_alloc_array_1d(jl_array_symbol_type, 0);
+    jl_array_t *a = jl_alloc_array_1d(jl_array_symbol_type, 0, 0);
     JL_GC_PUSH1(&a);
     size_t i;
     void **table = m->bindings.table;