--- a/test/core.jl
+++ b/test/core.jl
@@ -4643,12 +4643,12 @@ end # module SOE
 # issue #15240
 @test_nowarn begin
     local p15240
-    p15240 = ccall(:jl_realloc, Ptr{Void}, (Ptr{Void}, Csize_t), C_NULL, 10)
-    ccall(:jl_free, Void, (Ptr{Void}, ), p15240)
+    p15240 = ccall(:jl_gc_realloc, Ptr{Void}, (Ptr{Void}, Csize_t), C_NULL, 10)
+    ccall(:jl_gc_free, Void, (Ptr{Void}, ), p15240)
 end
 
 # issue #19963
-@test_nowarn ccall(:jl_free, Void, (Ptr{Void}, ), C_NULL)
+@test_nowarn ccall(:jl_gc_free, Void, (Ptr{Void}, ), C_NULL)
 
 # Wrong string size on 64bits for large string.
 if Sys.WORD_SIZE == 64