This resolves #130.  The main goal is to use calloc when creating large arrays of zeros (only opting in to this behavior for appropriate eltypes).  But I had to make a number of other changes along the way:

- Small arrays of zeros are allocated as before, but zeroed using `memset`
- Renamed `jl_malloc` (and friends) to `jl_gc_malloc` to disambiguate from `jl_malloc_aligned`, which allocates memory that isn't gc counted
- Added `jl_gc_malloc_aligned` (and friends)
- Removed `jl_gc_managed_malloc`, the old gc counted, aligned `malloc`, which didn't have a matching `calloc`
- Freed up a bit in `jl_array_flags_t` by tracking whether the memory is aligned or not in `mallocarray_t` instead

Before I submit this, I'd still like to do the following:

- [ ] Check that these changes do indeed make things faster! (need to check both large and small arrays)
- [ ] Add special casing to `jl_gc_malloc_aligned` and `jl_gc_calloc_aligned`
- [ ] Add checks on the alignment argument (?)
- [ ] When using `jl_gc_realloc_aligned`, be smarter and only `memcpy` the memory being used by the array
- [ ] Double check the `zeros` tests
- [ ] Use calloc for more eltypes