@danielmatz 2017-03-11
I know I need to update this branch to fix the conflicts, but I have also been meaning to post some benchmark results.  Let me do the latter now.  I'll find time soon to fix the conflicts.  I also need to figure out why the tests fail on systems other than macOS.

I benchmarked both `zeros(Float64, s)` and `Array(Float64, s)` for a variety of sizes.  The 2048 and 2049 sizes are right at the breakpoint in `_new_array_` where it switches from allocating the buffer inline with the array or separately with `malloc` or `calloc`.

I'm comparing master to the point in my PR after I've switched array creation to use `jl_gc_malloc` and then to the full PR which uses `calloc` for `zeros`.

`Array(Float64, s)`:

| size | master (min/median) | gc malloc change (min/median) | full pr (min/median) |
| --- | --- | --- | --- |
| 10 | 110 ns / 120 ns | 110 ns / 130 ns | 110 ns / 120 ns |
| 2048 | 240 ns / 1.7 µs | 250 ns / 1.4 µs | 240 ns / 1.3 µs |
| 2049 | 240 ns / 1.5 µs | 270 ns / 2.1 µs | 390 ns / 2.1 µs |
| 10000 | 1.4 µs / 5.5 µs | 600 ns / 2.4 µs | 760 ns / 2.4 µs |
| 100000 | 7.3 µs / 8.7 µs | 3.5 µs / 3.9 µs | 3.6 µs / 3.9 µs |
| 1000000 | 300 µs / 340 µs | 1.8 µs / 4.1 µs | 2.4 µs / 4.1 µs |

`zeros(Float64, s)`:

| size | master (min/median) | gc malloc change (min/median) | full pr (min/median) |
| --- | --- | --- | --- |
| 10 | 42 ns / 56 ns | 43 ns / 110 ns | 38 ns / 100 ns |
| 2048 | 1.2 µs / 1.9 µs | 1.8 µs / 7.6 µs | 1.9 µs / 7.5 µs |
| 2049 | 1.5 µs / 2.8 µs | 2.2 µs / 7.5 µs | 2.3 µs / 7.5 µs |
| 10000 | 6.1 µs / 13 µs | 10 µs / 34 µs | 9.9 µs / 34 µs |
| 100000 | 91 µs / 230 µs | 100 µs / 370 µs | 3.4 µs / 3.6 µs |
| 1000000 | 1.0 ms / 2.2 ms | 1.1 ms / 3.6 ms | 3.6 µs / 4.0 µs |

Did you all expect the gc malloc change to so drastically improve the times for large arrays?

Is the slight degradation for small malloced arrays (size = 2049) OK?

The good news is that using `calloc` does help a lot for large arrays of zeros.