@ViralBShah 2017-03-12
Another good thing to check on is the performance of the shootout and the micro benchmarks. The slight degradation for 2049 appears ok to me for all the other gains.