@danielmatz 2017-03-21
Here are the benchmark results:

```julia
(Any["micro", "quicksort"], TrialJudgement(+0.17% => invariant))
(Any["micro", "parseint"], TrialJudgement(-1.27% => invariant))
(Any["micro", "randmatstat"], TrialJudgement(+1.40% => invariant))
(Any["micro", "pisum"], TrialJudgement(+0.02% => invariant))
(Any["micro", "fib"], TrialJudgement(-0.73% => invariant))
(Any["micro", "randmatmul"], TrialJudgement(+9.96% => invariant))
(Any["micro", "mandel"], TrialJudgement(+0.42% => invariant))
(Any["shootout", "k_nucleotide"], TrialJudgement(+4.76% => invariant))
(Any["shootout", "fasta"], TrialJudgement(+1.26% => invariant))
(Any["shootout", "spectralnorm"], TrialJudgement(-0.00% => invariant))
(Any["shootout", "revcomp"], TrialJudgement(+1.61% => invariant))
(Any["shootout", "pidigits"], TrialJudgement(+49.25% => regression))
(Any["shootout", "binary_trees"], TrialJudgement(+11.79% => invariant))
(Any["shootout", "nbody"], TrialJudgement(+0.01% => invariant))
(Any["shootout", "regex_dna"], TrialJudgement(+8.55% => invariant))
(Any["shootout", "meteor_contest"], TrialJudgement(+7.98% => invariant))
(Any["shootout", "mandelbrot"], TrialJudgement(+0.28% => invariant))
(Any["shootout", "fannkuch"], TrialJudgement(-0.63% => invariant))
(Any["shootout", "nbody_vec"], TrialJudgement(+0.10% => invariant))
```

The pidigits regression is caused by spending a lot more time in the `BigInt()` constructor, which is getting called by `<<(::BigInt, ::UInt64)`.  I'm not sure yet why my changes would cause that...