@JeffBezanson 2017-07-25
Looks good. I'm going to rebase and take all but the last two commits, and implement #9147 instead, which will be simpler since we won't need the extra argument to request zeros.