@JeffBezanson 2017-07-25
@danielmatz Please review #22953 if you get a chance. I think that part should be ok to merge and you can submit future PRs to do the follow-up items if you want.